Updated MemberWrapper for YAXNamespaceAttribute
Should make it possible for MemberWrapper to detect and handle the
YAXNamespaceAttribute attribute when applied to fields or properties.
The difference between the implementation here and on UdtWrapper is that
this will throw an exception if a default namespace is applied to a
child element. This is to prevent children from breaking the namespace
of the parent document.