Updated MemberWrapper for YAXNamespaceAttribute
Made it more clear that the parent class/struct/whatever's default
namespace will be inherited automatically if no namespace is specified
for a child element.
Currently, explicitly redefining the child element's default namespace
will throw an exception (even if it is the same as the root document's
namespace) for safety's sake.