Added MultipleNamespaceSerializationTest
Should correctly serialize a document which specifies multiple different
namespaces, one for the root element (a custom namespace just to be
sure), an inherited parent namespace for one property, and two custom
namespaces for the other properties.
The idea is that the first property [BoolItem] will inherit the parent
document's namespace, while the other two will add namespaces to the
parent document and use those.